replace class attributes by instance attributes
To make schema definition easier, we use class attributes, so as to mimic a
declarative language. This has the consequence however that when a schema
modifies an attribute it inherited, it effectively modifies all other schema
instances that also inherited from its parent schema.

To avoid this, we replace all class attributes by instance attributes upon
instantiation.